+++ primint/src/num.rs
@@ -1,0 +1,5 @@
+//! Mirrors the [`core::num`] module.
+
+mod nonzero;
+
+pub use nonzero::NonZero;